IRInterpreter: fix sign extension of small types (pr37840)
Sign-extension of small types (e.g. short) was not handled correctly.
The reason for that was that when we were assigning the a value to the
Scalar object, we would accidentally promote the type to int (even
though the assignment code in AssignTypeToMatch tried to cast the value
to the appropriate type, it would still invoke the "int" version of
operator=). Instead, I use the APInt version of operator=, where the
bitwidth is specified explicitly. Among other things, this allows us to
fold the individual size cases into one.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@335114 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 48578b5aa03ef4a4d575dd76d7a985f9bcc60834
apple-llvm-split-dir: lldb/