[tsan] Avoid calling Block_copy in the "sync" GCD interceptors
The synchronous dispatch functions in GCD (dispatch_sync, dispatch_barrier_sync), don't make a copy of the passed block. To maintain binary compatibility, we should avoid doing that as well in TSan, as there's no reason to do that. The synchronous dispatch functions will not return before the block is actually executed.

rdar://problem/42242579

Differential Revision: https://reviews.llvm.org/D50920



git-svn-id: https://llvm.org/svn/llvm-project/compiler-rt/trunk@340342 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 3900ec3f3d0039a04fa942112987b04ebeaa262d
apple-llvm-split-dir: compiler-rt/